fix(bin): replaced index.js with cli.js to avoid exeuction of importing
this allows importing types and utils from @branchlint/cli
