+++ .changeset/breezy-moose-hear.md
@@ -1,0 +1,5 @@
+---
+"@branchlint/cli": patch
+---
+
+Changed bin to cli.js to avoid execute on import of types or utils
